code working
