+++ Models/AircraftManager.cs
@@ -7,17 +7,14 @@ namespace DTIWindow.Models
     {
         private static AircraftManager? _instance;
         public static AircraftManager Instance => _instance ??= new AircraftManager();
-        public BindingList<Aircraft> aircraftList = new BindingList<Aircraft>(); // List of aircraft in the system
-        public BindingList<Aircraft> AircraftList { get; private set; }
+        public BindingList<Aircraft> aircraftList = new BindingList<Aircraft>(); // Initialize here
+        public static BindingList<Aircraft> AircraftList => Instance.aircraftList; // Static property
         private static int nextAircraftNumber = 1; // Counter for generating unique aircraft names
-
-        public AircraftManager()
-        {
-            AircraftList = new BindingList<Aircraft>();
-        }
 
         public Aircraft GetOrCreateAircraft(string callsign)
         {
+            Debug.WriteLine($"GetOrCreateAircraft called with callsign: {callsign}");
+
             // Find the aircraft by callsign
             Aircraft? aircraft = AircraftList.FirstOrDefault(a => a.Callsign == callsign);
 
@@ -26,9 +23,12 @@ namespace DTIWindow.Models
                 // If the aircraft doesn't exist, create it
                 aircraft = new Aircraft($"Aircraft{nextAircraftNumber++}", callsign);
                 AircraftList.Add(aircraft); // Add to the shared list
+                Debug.WriteLine($"Created new aircraft: {callsign}");
             }
-
-            Debug.WriteLine($"Aircraft created or retrieved: {callsign}");
+            else
+            {
+                Debug.WriteLine($"Found existing aircraft: {callsign}");
+            }
 
             return aircraft;
         }
