+++ Models/Parents.cs
@@ -6,9 +6,8 @@ namespace DTIWindow.Models
     {
         public string Name { get; set; } // Name of the aircraft
         public string Callsign { get; set; } // Callsign of the aircraft
-        public BindingList<ChildAircraft> Children { get; set; } // List of child aircraft associated with this aircraft
+        public BindingList<ChildAircraft> Children { get; set; } = new BindingList<ChildAircraft>(); // List of child aircraft associated with this aircraft
         private Aircraft? designatedAircraft = null; // Currently designated aircraft
-        private BindingList<Aircraft> aircraftList = new BindingList<Aircraft>(); // List of aircraft in the system
 
         public Aircraft(string name, string callsign)
         {
@@ -21,6 +20,7 @@ namespace DTIWindow.Models
         // Adds a child aircraft to the list if it doesn't already exist
         public void AddChild(ChildAircraft child)
         {
+            // Check if a child with the same callsign already exists
             if (!Children.Any(c => c.Callsign == child.Callsign))
             {
                 Children.Add(child);
@@ -36,9 +36,11 @@ namespace DTIWindow.Models
         {
             designatedAircraft = aircraft;
 
-            // Refresh the UI to reflect the change
-            var windowInstance = new DTIWindow.UI.Window(aircraftList, new Dictionary<Aircraft, List<Aircraft>>());
-            windowInstance.PopulateAircraftDisplay();
+            // Refresh the existing Window instance
+            var windowInstance = Application.OpenForms
+                .OfType<Form>()
+                .FirstOrDefault(form => form.Name == "Window");
+            windowInstance?.GetType().GetMethod("PopulateAircraftDisplay")?.Invoke(windowInstance, null);
         }
     }
 }
