+++ UI/Colours.cs
@@ -1,5 +1,3 @@
-using System.Diagnostics;
-
 namespace DTIWindow.UI
 {
     public static class Colours
@@ -33,11 +31,8 @@ namespace DTIWindow.UI
         {
             try
             {
-                Debug.WriteLine($"GetHMIStateAndColor called with hmiState: {hmiState}");
-
                 if (string.IsNullOrEmpty(hmiState) || hmiState == "Unknown State")
                 {
-                    Debug.WriteLine("HMI state is null, empty, or 'Unknown State'. Returning default values.");
                     return ("Unknown State", Color.Gray); // Default for unknown states
                 }
 
@@ -55,18 +50,13 @@ namespace DTIWindow.UI
                     _ => vatsys.Colours.Identities.Default
                 };
 
-                Debug.WriteLine($"Mapped HMI state '{hmiState}' to vatsys.Colours identity: {identity}");
-
                 // Get the color from vatsys.Colours
                 Color color = vatsys.Colours.GetColour(identity);
 
-                Debug.WriteLine($"Retrieved color for identity '{identity}': {color}");
-
                 return (hmiState, color);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                Debug.WriteLine($"Error in GetHMIStateAndColor: {ex.Message}");
                 return ("Error", Color.Red); // Default error case
             }
         }
